styling profile page and related components
