+++ src/components/footer/Footer.jsx
@@ -15,6 +15,9 @@ const Footer = () => (
       justifyContent: 'space-around',
       backgroundColor: '#E4DCCF',
       height: '160px',
+      position: 'absolute',
+      bottom: 0,
+      width: '100%',
 
     }}
   >
