+++ src/components/menu/Menu.jsx
@@ -18,9 +18,6 @@ const MenuComp = () => {
     <div>
       <Button
         id="basic-button"
-        // aria-controls={open ? 'basic-menu' : undefined}
-        // aria-haspopup="true"
-        // aria-expanded={open ? 'true' : undefined}
         onClick={handleClick}
       >
         <PersonIcon />
