+++ src/components/table/index.jsx
@@ -3,7 +3,7 @@ import Box from '@mui/material/Box';
 import { DataGrid } from '@mui/x-data-grid';
 
 const Table = ({ rows, columns }) => (
-  <Box>
+  <Box m="auto" width="fit-content">
     <DataGrid
       rows={rows}
       columns={columns}
