+++ src/index.css
@@ -1,5 +1,9 @@
 * {
-  padding:0;
+  padding: 0;
   margin: 0;
   box-sizing: border-box;
+}
+
+body {
+  background-color: #F9F5EB;
 }
